Retool to the comparison to improve performance.
Rather than perform the cell expansion on each candidate test
structure (B' in the manuscript) and wrap the atoms from A into the B'
unit cell during comparison, the new code expands description A in the
preparation stage and then wraps the B' atoms into A's cell during
comparison.

This saves quite a few coordinate conversions, and reduces the number
of expansions (the most expensive portion of the algorithm according
to the code profiles I've generated) from >16 to 1.

I reran my benchmarks (is XtalOpt's xtaltest) where I generate a
random structure (call it desc1) and then randomize the description
while preserving the structure (to create desc2). I then compare them
to ensure that they match. I then randomly displace one of the atoms
in desc2 by a distance > the tolerance (to change the structure),
compare desc1 and desc2 again to ensure that the comparison fails.

I time both of the comparisons, since the negative comparison should
be much slower than the positive test. For the both the old and new
code's successes and failures, I get the timings below. The numbers
are averages of 2500 comparisons of structures with 100 atoms each.
Units are ms/comparison:

Old code:
positive comp: 0.686
negative comp: 2.472

New code:
positive comp: 0.691
negative comp: 1.003

So not much improvement when the structures match, but a ~60% decrease
in time when they don't.