Cleanup trailing whitespace, indentation.
Trailing whitespace creates a lot of noise when looking at diffs,
and some of developers (including me) are OCD about it. Most text editors
have the ability to clean-up whitespace on saving (preferably only on
modified lines) so that's a good idea when you work on this stuff a lot.

Extra spaces between blocks are another OCD thing, and they also lessen
the amount of code that fits on screen at once.

Indentation is just a readability thing.

I also removed a pointless comment I'd made in xtalcomp.h to keep a line
under 80 characters, since it was just a leftover from when I was
developing this.