Use a map<origIdx, pair<dupFirstIdx, dupLastIdx> > for duplicates.
This uses a more appropriate type for the duplicate lookup, and adjusted
a few bits to work with the std::map API.

I removed the goto's -- these are not really used anymore. They have
weird scoping rules, break up the flow of the program (readability-wise),
and get really nasty when overused. I shortened the insertion code using
map::insert, and just inlined this where the goto's were.