Update spglib primitive reduction call
This seems to be the proper "new" way to perform a
primitive reduction with spglib. The old way that we
were using it would sometimes cause strange results
(such as finding a match with one ordering and not finding
a match with a different ordering).

That bug seems to go away when we call the primitive reducing
functions like the spglib documentation suggests.